moves api to users.js, restores index.js
